Set different start time for live streaming Add external link to the sabbath school
Note: bypassSecurityTrustHtml is meant to mark safe values, which are not susceptible to XSS attacks.
In our case the whole URL to the sabbath school is hardcoded, thus making it safe in nature.
This clarification is needed, because the method name can leave wrong impression,
that we're actually bypassing something inherently insecure, which is not true